refactor: move tools unit tests to tools/__tests__
